Module lzw::bitstream

This module provides bit readers and writers

Macros

define_bit_readers

define_bit_writers

Structs

LsbReader

Reads bits from a byte stream, LSB first.

LsbWriter

Writes bits to a byte stream, LSB first.

MsbReader

Reads bits from a byte stream, MSB first.

MsbWriter

Writes bits to a byte stream, MSB first.

Enums

Bits

Containes either the consumed bytes and reconstructed bits or only the consumed bytes if the supplied buffer was not bit enough

Traits

BitReader

A bit reader.

BitWriter

A bit writer.