Enum lzw::Bits

pub enum Bits {
    Some(usize, u16),
    None(usize),
}

Containes either the consumed bytes and reconstructed bits or only the consumed bytes if the supplied buffer was not bit enough

Variants
Some(usize, u16)

Consumed bytes, reconstructed bits

None(usize)

Consumed bytes