Struct lzw::lzw::Decoder

pub struct Decoder<R: BitReader> {
    r: R,
    prev: Option<Code>,
    table: DecodingDict,
    buf: [u8; 1],
    code_size: u8,
    min_code_size: u8,
    clear_code: Code,
    end_code: Code,
}

Decoder for a LZW compressed stream (this algorithm is used for GIF files).

The maximum supported code size is 16 bits. The decoder assumes two special code word to be present in the stream:

CLEAR_CODE == 1 << min_code_size

END_CODE == CLEAR_CODE + 1

Furthermore the decoder expects the stream to start with a CLEAR_CODE. This corresponds to the implementation needed for en- and decoding GIF and TIFF files.

Fields

r

prev

table

buf

code_size

min_code_size

clear_code

end_code

Methods

impl<R> Decoder<R> where R: BitReader

fn new(reader: R, min_code_size: u8) -> Decoder<R>

Creates a new LZW decoder.

fn decode_bytes(&mut self, bytes: &[u8]) -> Result<(usize, &[u8])>

Tries to obtain and decode a code word from bytes.

Returns the number of bytes that have been consumed from bytes. An empty slice does not indicate EOF.